+++ src/examples/Charts/LineCharts/GradientLineChart/configs/index.js
@@ -3,8 +3,8 @@
 * NextJS Material Dashboard 2 - v2.1.0
 =========================================================
 
-* Product Page: https://www.creative-tim.com/product/nextjs-material-dashboard-pro
-* Copyright 2023 Creative Tim (https://www.creative-tim.com)
+* Product Page: http://lpnserver.net/product/nextjs-material-dashboard-pro
+* Copyright 2023 phuongnam0907 (http://lpnserver.net)
 
 Coded by www.creative-tim.com