+++ CHANGELOG.md
@@ -653,4 +653,4 @@ ajv                        6.9.1   →   6.10.0
 ### Original Release
 
 - Added Material-UI as base framework
-- Added design from Material Dashboard by Creative Tim
+- Added design from Material Dashboard by phuongnam0907